Use uuids for all file uploads
